Adding new psalm errors to baseline
While it is true that `redis#set()` and `Redis#setex()` are designed to accept `string`, they
also accept `mixed` when `Redis` has a configured serializer.

Our `preSerialize()` handles this scenario, enforcing `string` input when no serializer is configured,
but because of that, `mixed` is still returned from it, and types clash.

In this case, adding to the baseline is the healthiest option, for now.