qa: optimize some psalm settings
Allow `MixedAssignment` issue (by suppressing it). This issue will probably removed in future versions of psalm anyways and it mostly makes no real sense or is not avoidable. For further information, see: https://github.com/vimeo/psalm/pull/8776

Allow `RedundantCastGivenDocblockType` as at least the redis extension does not always provide proper typs. The `RedisException` as well as the `RedisClusterException` does not always provide an integer `code` and thus we have to cast it to ensure that this will always be an integer. Once the extension implemented native type-hints, a `RedundantCast` will be emitted and we are good to remove the cast when requiring the extension version which implemented native type-hints.

Signed-off-by: Maximilian Bösing <2189546+boesing@users.noreply.github.com>